Improved Hyperlink Formatting (Underline)
Activated hidelinks option for those who don't want visible links and added improved formatting for hyperlinks such that they are now underlined instead of having wonky boxes. The underlined setting overwrites the hidelinks option so long as it is not commented out and the underlines are not printed out to paper.